# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Amir Numerique - Digital Printing Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all active services
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /services/slug/{slug}:
    get:
      operationId: getServiceBySlug
      tags: [services]
      summary: Get service by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List quote requests (admin gets all, user gets own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteListResponse"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get quote by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuoteStatus
      tags: [quotes]
      summary: Update quote status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteStatusBody"
      responses:
        "200":
          description: Quote updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/profile:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /settings/{key}:
    put:
      operationId: updateSetting
      tags: [settings]
      summary: Update a setting (admin only)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingBody"
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [visitor, client, subcontractor, admin]
        preferredLanguage:
          type: string
          enum: [fr, ar]
        accountStatus:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - preferredLanguage
        - accountStatus
        - createdAt
...
    Service:
      type: object
      properties:
        id:
          type: integer
        nameFr:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        descriptionFr:
          type: string
          nullable: true
        descriptionAr:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        publicPricePerM2:
          type: number
        clientPricePerM2:
          type: number
        subcontractorPricePerM2:
          type: number
        requiresFileUpload:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameFr
        - nameAr
        - slug
        - publicPricePerM2
        - clientPricePerM2
        - subcontractorPricePerM2
        - requiresFileUpload
        - active
        - createdAt
    CreateServiceBody:
      type: object
      properties:
        nameFr:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        publicPricePerM2:
          type: number
        clientPricePerM2:
          type: number
        subcontractorPricePerM2:
          type: number
        requiresFileUpload:
          type: boolean
          default: false
        active:
          type: boolean
          default: true
      required:
        - nameFr
        - nameAr
        - slug
        - publicPricePerM2
        - clientPricePerM2
        - subcontractorPricePerM2
    UpdateServiceBody:
      type: object
      properties:
        nameFr:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        descriptionFr:
          type: string
        descriptionAr:
          type: string
        imageUrl:
          type: string
        publicPricePerM2:
          type: number
        clientPricePerM2:
          type: number
        subcontractorPricePerM2:
          type: number
        requiresFileUpload:
          type: boolean
        active:
          type: boolean
...
    Quote:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        widthInput:
          type: number
        heightInput:
          type: number
        unitInput:
          type: string
          enum: [cm, m]
        widthM:
          type: number
        heightM:
          type: number
        areaM2:
          type: number
        estimatedPrice:
          type: number
          nullable: true
        note:
          type: string
          nullable: true
        fileUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, responded, accepted, refused, converted_to_order]
        adminResponse:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        service:
          $ref: "#/components/schemas/Service"
          nullable: true
      required:
        - id
        - userId
        - serviceId
        - widthInput
        - heightInput
        - unitInput
        - widthM
        - heightM
        - areaM2
        - status
        - createdAt
        - updatedAt
    CreateQuoteBody:
      type: object
      properties:
        serviceId:
          type: integer
        widthInput:
          type: number
        heightInput:
          type: number
        unitInput:
          type: string
          enum: [cm, m]
        note:
          type: string
        fileUrl:
          type: string
      required:
        - serviceId
        - widthInput
        - heightInput
        - unitInput
    UpdateQuoteStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, responded, accepted, refused, converted_to_order]
        adminResponse:
          type: string
      required:
        - status
    QuoteListResponse:
      type: object
      properties:
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - quotes
        - total
        - page
        - limit
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [client, subcontractor, admin]
        accountStatus:
          type: string
          enum: [active, suspended]
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        preferredLanguage:
          type: string
          enum: [fr, ar]
        currentPassword:
          type: string
        newPassword:
          type: string
    Setting:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
      required:
        - id
        - key
        - value
    UpdateSettingBody:
      type: object
      properties:
        value:
          type: string
      required:
        - value
...