# lib/api-zod/src/generated/types/updateServiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amir Numerique - Digital Printing Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServiceBody {
  nameFr?: string;
  nameAr?: string;
  slug?: string;
  descriptionFr?: string;
  descriptionAr?: string;
  imageUrl?: string;
  publicPricePerM2?: number;
  clientPricePerM2?: number;
  subcontractorPricePerM2?: number;
  requiresFileUpload?: boolean;
  active?: boolean;
}