# lib/api-zod/src/generated/types/updateSettingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Amir Numerique - Digital Printing Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSettingBody {
  value: string;
}